refactor(tmux): extract session to its own file
